requirements.txt
asgiref==3.2.10
coverage==5.2.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-adminlte2==0.4.3
django-stdimage==5.1.1
gunicorn==20.0.4
model-mommy==2.0.0
Pillow==7.2.0
progressbar2==3.51.4
psycopg2-binary==2.8.5
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
